Refactored folders and modified css for map
