+++ frontend/sortify/app/component/header/header.module.css
@@ -56,7 +56,7 @@
     background-color: #87C75C; /* Green background for the button */
     color: #0B540D;
     padding: 12px 24px;
-    margin-right: 8px 0; /* Add some vertical space between items */
+    /*margin-right: 8px 0; !* Add some vertical space between items *!*/
     border-radius: 30px; /* Rounded edges */
     text-decoration: none;
     font-weight: 600; /* Slightly bold for emphasis */
@@ -80,7 +80,7 @@
     top: 100%; /* Position directly below the button */
     left: 0;
     background-color: black; /* Green background for the menu */
-    box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.1);
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
     border-radius: 8px;
     min-width: 160px;
     z-index: 2000;
