--- frontend/sortify/app/globals.css
@@ -1,35 +1,0 @@
-html,
-body {
-  margin: 0;
-  padding: 0;
-  width: 100%;
-  height: 100%;
-  font-family: "Poppins", sans-serif;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  background-color: #CFD1CC;
-  justify-content: space-between;
-}
-
-.main-content {
-  width: 100%;
-  display: flex;
-  margin-top: 2%;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  text-align: center;
-}
-
-#map {
-  border-radius: 15px;      /* Rounded corners */
-  margin-top: 20px;
-  margin-bottom: 40px;
-  box-sizing: border-box;
-  overflow: hidden;
-  box-shadow: 0 0 0 2px #545454;
-}
-
-
