+++ frontend/sortify/app/layout.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import Header from "./component/header"
+import "./global.css";
+import Header from "./component/header/header"
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v15-appRouter';
-import Footer from "./component/footer";
+import Footer from "./component/footer/footer";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -28,6 +28,7 @@ export default function RootLayout({
   return (
     <html lang="en">
     <head>
+      <title></title>
     </head>
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
         <AppRouterCacheProvider>
